backend/requirements.txt
Django==5.2.2
djangorestframework
django-cors-headers
requests
python-decouple